+++ src/main/kotlin/services/queue/GuildQueueService.kt
@@ -3,10 +3,15 @@ package es.wokis.services.queue
 import dev.kord.common.entity.Snowflake
 import dev.kord.core.behavior.channel.BaseVoiceChannelBehavior
 import dev.kord.core.entity.channel.MessageChannel
+import dev.kord.core.entity.interaction.ApplicationCommandInteraction
 import es.wokis.dispatchers.AppDispatchers
+import es.wokis.localization.LocalizationKeys
 import es.wokis.services.lavaplayer.AudioPlayerManagerProvider
 import es.wokis.services.lavaplayer.GuildLavaPlayerService
 import es.wokis.services.localization.LocalizationService
+import es.wokis.utils.getMemberVoiceChannel
+import es.wokis.utils.orDefaultLocale
+import kotlin.jvm.Throws
 
 class GuildQueueService(
     private val audioPlayerManagerProvider: AudioPlayerManagerProvider,
@@ -15,6 +20,23 @@ class GuildQueueService(
 ) {
 
     private val guildQueues: MutableMap<Snowflake, GuildLavaPlayerService> = mutableMapOf()
+
+    @Throws(IllegalStateException::class)
+    suspend fun getOrCreateLavaPlayerService(interaction: ApplicationCommandInteraction): GuildLavaPlayerService {
+        val locale = interaction.guildLocale.orDefaultLocale()
+        val voiceChannel = interaction.getMemberVoiceChannel(interaction.kord)
+            ?: throw IllegalStateException(localizationService.getString(LocalizationKeys.ERROR_NO_VOICE_CHANNEL, locale))
+        val textChannel = interaction.channel.asChannelOrNull()
+            ?: throw IllegalStateException(localizationService.getString(LocalizationKeys.ERROR_NO_TEXT_CHANNEL, locale))
+        val guildId = interaction.data.guildId.value
+            ?: throw IllegalStateException(localizationService.getString(LocalizationKeys.ERROR_NO_GUILD, locale))
+
+        return getOrCreateLavaPlayerService(
+            guildId = guildId,
+            textChannel = textChannel,
+            voiceChannel = voiceChannel
+        )
+    }
 
     fun getOrCreateLavaPlayerService(
         guildId: Snowflake,
