+++ src/main/kotlin/utils/InteractionExtensions.kt
@@ -2,9 +2,10 @@ package es.wokis.utils
 
 import dev.kord.core.Kord
 import dev.kord.core.behavior.channel.BaseVoiceChannelBehavior
+import dev.kord.core.entity.interaction.ApplicationCommandInteraction
 import dev.kord.core.entity.interaction.ChatInputCommandInteraction
 
-suspend fun ChatInputCommandInteraction.getMemberVoiceChannel(bot: Kord): BaseVoiceChannelBehavior? = data.guildId.value?.let { guildId ->
+suspend fun ApplicationCommandInteraction.getMemberVoiceChannel(bot: Kord): BaseVoiceChannelBehavior? = data.guildId.value?.let { guildId ->
     bot.getGuildOrNull(guildId)?.let { guild ->
         guild.getMemberOrNull(user.id)?.let { member ->
             member.getVoiceStateOrNull()?.getChannelOrNull()
